Add comprehensive README and examples, fix linter issues
- Update README.md with complete feature documentation including:
  - Enterprise security features (MFA, session management, rate limiting)
  - LDAP/AD integration with group sync
  - Certificate management with HSM support
  - Comprehensive audit logging
  - Health checks and Prometheus metrics
  - All API endpoints and configuration options

- Add detailed examples for common use cases:
  - OAuth2 client credentials flow (Python, Node.js, Go, Rust)
  - OAuth2 authorization code flow with PKCE (React, Flask)
  - OpenID Connect integration
  - API key management for CI/CD

- Fix clippy linter issues:
  - Remove unused imports across multiple modules
  - Fix collapsible if statements using let chains
  - Replace manual modulo checks with .is_multiple_of()
  - Remove unnecessary type casts
  - Fix empty line after doc comments
  - Add #[allow(clippy::type_complexity)] for complex return types
  - Prefix unused parameters with underscore in stub implementations

- Test results: 89 passed, 5 failed (certificate module tests)
  - Most tests passing, remaining failures in cert rotation tests

+++ src/middleware/rate_limit_redis.rs
@@ -141,7 +141,7 @@ impl RateLimiter for RedisRateLimiter {
         if attempts >= self.max_failed_attempts {
             debug!("Blocking {} after {} failed attempts", key, attempts);
             let _: () = conn
-                .set_ex(&block_key, 1, self.block_duration_secs as u64)
+                .set_ex(&block_key, 1, self.block_duration_secs)
                 .await
                 .map_err(|e| format!("Redis error: {}", e))?;
         }
